requirements.txt
flask
Flask-session
requests
mysql.connector
cs50
cs50
Flask
requests
